pyproject.toml
[tool.poetry]
name = "mdx-to-api"
version = "0.1.0"
description = ""
authors = ["bigzhu <vermiliondun@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
uvicorn = { extras = ["standard"], version = "^0.30.1" }
pyjwt = "^2.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pyright]
venvPath = "."
venv = ".venv"
